Add in Shared function for Trimming whitespace in plates
This PR is the foundation for Trimming plates in all resources that reference vehicle plates, this will allow plates with less than 8 Characters to work with QB resources (mechanic and garage mainly), and not return white spaces. For example, the plate "TEST" would return "  TEST  " so when trying to put it back into the garage or add mods to it, it would fail.

For example in qb-garages: TriggerEvent("vehiclekeys:client:SetOwner", QBCore.Functions.GetPlate(veh)) would return "TEST instead of "  TEST  " when using the native.

What I did was replace GetVehicleNumberPlateText( with QBCore.Functions.GetPlate( in all resources that contain the QBCore object.