Spawning vehicles through the server
In writing these posts, I relied on Spawning A Vehicle At A Distance in discord server
The following notes through the words of Mr.Kakarot

This will spawn the entity and callback the entity id which you can then use on the client side to check if the entity exists. It will NOT exist until a client(player) gets within range of it! While a client is NOT in range, the server “owns” the entity. So whenever a client gets into infinity range it then takes ownership of that entity and can be used by them.