New TransferMoney function and fix
New TransferMoney function, using mysql transactions, this should be used instead of Remove and AddMoney, sice this locks the table until the transaction is complete.

A fix as well for PlayerData.online, since it gave errors sometimes, now it only uses numbers, and not a mix of booleans and numbers.